+++ integrators/euler/euler.h
@@ -15,12 +15,12 @@ public:
 	gpu_euler_integrator(const config &cfg);
 
 public:
-	void integrate(gpu_ensemble &ens, float T);
+	void integrate(gpu_ensemble &ens, double T);
 
 	// No support for CPU execution. Note: we could make this function
 	// transparently copy the CPU ensemble to the GPU, and back once
 	// the integration is done
-	void integrate(cpu_ensemble &ens, float T) { abort(); }
+	void integrate(cpu_ensemble &ens, double T) { abort(); }
 };
 
 #endif
